# apps/api/swagger.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/organizations/{organizationId}/services": {
      "post": {
        "description": "Create a new service within an organization.",
        "operationId": "CreateServiceController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateServiceBodyDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response with created service details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateServiceResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Create a new service",
        "tags": ["Services"]
      },
      "get": {
        "description": "Retrieve all services within an organization.",
        "operationId": "GetAllServicesController",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number for pagination",
            "schema": { "type": "number" }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "Number of items per page for pagination",
            "schema": { "type": "number" }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "Order of the results (asc or desc)",
            "schema": { "enum": ["asc", "desc"], "type": "string" }
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "description": "Field to order the results by",
            "schema": { "enum": ["name", "createdAt", "updatedAt"], "type": "string" }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term to filter services by name",
            "schema": { "type": "string" }
          },
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with all services details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetAllServicesResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "summary": "Get all services",
        "tags": ["Services"]
      }
    },
    "/organizations/{organizationId}/services/{serviceId}": {
      "delete": {
        "description": "Delete a service within an organization.",
        "operationId": "DeleteServiceController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the service to be deleted",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with deleted service details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteServiceResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Delete a service",
        "tags": ["Services"]
      },
      "put": {
        "description": "Update the details of a service within an organization.",
        "operationId": "UpdateServiceController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the service",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateServiceBodyDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response with updated service details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UpdateServiceResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Update a service",
        "tags": ["Services"]
      }
    },
...
    "/organizations/{organizationId}/services/{serviceId}/service-staff": {
      "get": {
        "description": "Retrieve a list of staff members assigned to a specific service within an organization.",
        "operationId": "GetServiceStaffByServiceIdController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          },
          {
            "name": "serviceId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the service",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with service staff details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetServiceStaffByServiceIdResponseDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get service staff by service ID",
        "tags": ["Service Staff"]
      }
    },
    "/organizations/{organizationId}/by-staff/service-staff": {
      "post": {
        "description": "Retrieve service staff members associated with a specific staff ID within an organization.",
        "operationId": "GetServiceStaffByStaffIdController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": {
              "format": "ulid",
              "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with service staff details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetServiceStaffByStaffIdResponseDto_Output"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/NotFoundErrorDto" } }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get service staff by staff ID",
        "tags": ["Service Staff"]
      }
    },
...
    "/organizations/{organizationId}/analytics": {
      "get": {
        "description": "Retrieve served tickets count and average service duration.",
        "operationId": "GetAnalyticsController",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "description": "The unique identifier of the organization",
            "schema": { "type": "string" }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "description": "Number of days to look back for served tickets (default: 7)",
            "schema": { "minimum": 1, "maximum": 365, "default": 7, "type": "number" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with analytics data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetAnalyticsResponseDto_Output" }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UnauthorizedErrorDto" }
              }
            }
          }
        },
        "security": [{ "bearer": [] }],
        "summary": "Get historical analytics",
        "tags": ["Analytics"]
      }
    },
...
  },
  "info": {
    "title": "Queue management",
    "description": "A queue management API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
      "NotFoundErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "const": 404 },
          "message": { "type": "string" }
        },
        "required": ["statusCode", "message"]
      },
...
      "UnauthorizedErrorDto": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "number", "const": 401 },
          "message": { "type": "string" }
        },
        "required": ["statusCode", "message"]
      },
...
      "CreateServiceBodyDto": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "description": { "type": "string" },
          "avgDurationInt": { "type": "number" },
          "isActive": { "type": "boolean" },
          "maxCapacity": { "type": "number" }
        },
        "required": ["name"]
      },
      "CreateServiceResponseDto_Output": {
        "type": "object",
        "properties": {
          "service": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "ulid",
                "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
              },
              "name": { "type": "string" },
              "description": { "type": "string" },
              "avgDurationInt": { "type": "number" },
              "maxCapacity": { "anyOf": [{ "type": "number" }, { "type": "null" }] },
              "alertThresholdMinutes": { "type": "number" },
              "isActive": { "type": "boolean" },
              "organizationId": { "type": "string" },
              "createdAt": { "type": "string" },
              "updatedAt": { "type": "string" }
            },
            "required": ["id", "name", "organizationId", "createdAt", "updatedAt"],
            "additionalProperties": false
          }
        },
        "required": ["service"],
        "additionalProperties": false
      },
      "DeleteServiceResponseDto_Output": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "required": ["message"],
        "additionalProperties": false
      },
      "GetAllServicesResponseDto_Output": {
        "type": "object",
        "properties": {
          "services": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "ulid",
                  "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
                },
                "name": { "type": "string" },
                "description": { "type": "string" },
                "avgDurationInt": { "type": "number" },
                "maxCapacity": { "anyOf": [{ "type": "number" }, { "type": "null" }] },
                "alertThresholdMinutes": { "type": "number" },
                "isActive": { "type": "boolean" },
                "organizationId": { "type": "string" },
                "createdAt": { "type": "string" },
                "updatedAt": { "type": "string" }
              },
              "required": ["id", "name", "organizationId", "createdAt", "updatedAt"],
              "additionalProperties": false
            }
          },
          "total": { "type": "number" }
        },
        "required": ["services", "total"],
        "additionalProperties": false
      },
...
      "UpdateServiceBodyDto": {
        "type": "object",
        "properties": {
          "isActive": { "type": "boolean" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "avgDurationInt": { "type": "number" }
        }
      },
      "UpdateServiceResponseDto_Output": {
        "type": "object",
        "properties": {
          "service": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "ulid",
                "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
              },
              "name": { "type": "string" },
              "description": { "type": "string" },
              "avgDurationInt": { "type": "number" },
              "maxCapacity": { "anyOf": [{ "type": "number" }, { "type": "null" }] },
              "alertThresholdMinutes": { "type": "number" },
              "isActive": { "type": "boolean" },
              "organizationId": { "type": "string" },
              "createdAt": { "type": "string" },
              "updatedAt": { "type": "string" }
            },
            "required": ["id", "name", "organizationId", "createdAt", "updatedAt"],
            "additionalProperties": false
          }
        },
        "required": ["service"],
        "additionalProperties": false
      },
...
      "GetServiceStaffByServiceIdResponseDto_Output": {
        "type": "object",
        "properties": {
          "servicesStaff": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "ulid",
                  "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
                },
                "userId": {
                  "type": "string",
                  "format": "ulid",
                  "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
                },
                "serviceId": {
                  "type": "string",
                  "format": "ulid",
                  "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
                },
                "active": { "type": "boolean" },
                "isOnline": { "type": "boolean" },
                "isCounterClosed": { "type": "boolean" }
              },
              "required": ["id", "userId", "serviceId"],
              "additionalProperties": false
            }
          }
        },
        "required": ["servicesStaff"],
        "additionalProperties": false
      },
      "GetServiceStaffByStaffIdResponseDto_Output": {
        "type": "object",
        "properties": {
          "servicesStaff": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "ulid",
                  "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
                },
                "userId": {
                  "type": "string",
                  "format": "ulid",
                  "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
                },
                "serviceId": {
                  "type": "string",
                  "format": "ulid",
                  "pattern": "^[0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{26}$"
                },
                "active": { "type": "boolean" },
                "isOnline": { "type": "boolean" },
                "isCounterClosed": { "type": "boolean" }
              },
              "required": ["id", "userId", "serviceId"],
              "additionalProperties": false
            }
          }
        },
        "required": ["servicesStaff"],
        "additionalProperties": false
      },
...
      "GetAnalyticsResponseDto_Output": {
        "type": "object",
        "properties": {
          "servedTicketsByDay": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": { "date": { "type": "string" }, "count": { "type": "number" } },
              "required": ["date", "count"],
              "additionalProperties": false
            }
          },
          "avgDurationByEmployee": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "employeeId": { "type": "string" },
                "employeeName": { "type": "string" },
                "avgDuration": { "type": "number" }
              },
              "required": ["employeeId", "employeeName", "avgDuration"],
              "additionalProperties": false
            }
          }
        },
        "required": ["servedTicketsByDay", "avgDurationByEmployee"],
        "additionalProperties": false
      },
...
    }
  }
}